retc: Update submodule pointers for reipset and libusrxml
Update submodule pointer for reipset to ee1c8fb70b6f.
Update submodule pointer for libusrxml to c52f1ae0d9fd.

With this submodule update prepare_file() function improved to accept
either absolute path to file to prepare that is relative to ROOT or
relative path that is prefixed with ROOT internally.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>